refa: created a new package for eva & move structure Identifier out
