+++ client/src/main/java/edu/pku/code2graph/client/evaluation/XLLEvaluation.java
@@ -1,7 +1,9 @@
-package edu.pku.code2graph.client;
+package edu.pku.code2graph.client.evaluation;
 
 import com.csvreader.CsvReader;
 import com.csvreader.CsvWriter;
+import edu.pku.code2graph.client.Code2Graph;
+import edu.pku.code2graph.client.MybatisPreprocesser;
 import edu.pku.code2graph.diff.RepoAnalyzer;
 import edu.pku.code2graph.diff.util.GitService;
 import edu.pku.code2graph.diff.util.GitServiceCGit;
